refactor : ThemeFlag 위치 변경
