+++ client/src/components/Broadcast/Broadcast.styles.tsx
@@ -1,6 +1,6 @@
 import styled from 'styled-components'
 import TYPO from '@/styles/typo/TYPO'
-import { ThemeFlag } from '@/states/theme'
+import { ThemeFlag } from '@/types/theme'
 
 interface BroadcastProps {
   index: number
