+++ client/src/components/Modal/ViewerModal/ViewerModal.stories.tsx
@@ -1,6 +1,6 @@
 import type { StoryObj } from '@storybook/react'
 import ViewerModal from './ViewerModal'
-import { ThemeFlag } from '@/states/theme'
+import { ThemeFlag } from '@/types/theme'
 
 const meta = {
   title: 'Modal/Viewer',
