+++ client/src/states/user.ts
@@ -1,5 +1,5 @@
-import { User } from '@/types/user'
 import { atom } from 'recoil'
+import { User } from '@/types/user'
 
 export const userState = atom<User>({
   key: 'USER_STATE',
