+++ client/src/types/theme.ts
@@ -1,4 +1,7 @@
-import { ThemeFlag } from '@/states/theme'
+export enum ThemeFlag {
+  light,
+  dark,
+}
 
 export interface ThemeInterface {
   currentTheme: ThemeFlag
